+++ index.js
@@ -2,25 +2,24 @@
 module.exports = gulpRequireTasks;
 
 
+const path = require('path');
+const requireDirectory = require('require-directory');
+
+
 const DEFAULT_OPTIONS = {
   path: process.cwd() + '/gulp-tasks',
   include: null,
   exclude: null,
   separator: ':',
-  arguments: [],
   passGulp: true,
   passCallback: true,
   gulp: null
 };
 
-const path = require('path');
-const requireDirectory = require('require-directory');
-const _ = require('lodash');
-
 
 function gulpRequireTasks (options) {
 
-  options = _.extend({}, DEFAULT_OPTIONS, options);
+  options = Object.assign({}, DEFAULT_OPTIONS, options);
 
   const gulp = options.gulp || require('gulp');
   const gulp4 = !!gulp.series;
@@ -33,6 +32,7 @@ function gulpRequireTasks (options) {
     exclude: options.exclude
   });
 
+
   /**
    * Registers the specified module. Task name is deducted from the specified path.
    *
@@ -44,21 +44,30 @@ function gulpRequireTasks (options) {
     var module = normalizeModule(module, gulp4);
     var batchName;
 
+    if (module.dep) {
+      console.warn(
+        'Usage of "module.dep" property is deprecated and will be removed in next major version. ' +
+        'Use "deps" instead.'
+      );
+    }
+
+    var deps = module.deps || module.dep
+
     if (gulp4) {
-      if (module.dep) {
+      if (deps) {
         // Can get dep: [] which is equal to dep: { series: [] }
         // or dep: { parallel: [] }
-        if (_.isPlainObject(module.dep)) {
-          batchName = module.dep.parallel ? 'parallel' : 'series';
-          args.push(gulp[batchName].apply(gulp, module.dep[batchName]));
+        if (_.isPlainObject(deps)) {
+          batchName = deps.parallel ? 'parallel' : 'series';
+          args.push(gulp[batchName].apply(gulp, deps[batchName]));
         } else {
-          args.push(gulp.series.apply(gulp, module.dep));
+          args.push(gulp.series.apply(gulp, deps));
         }
       } else {
         args.push(module.nativeTask || taskFunction);
       }
     } else {
-      args.push(module.dep, module.nativeTask || taskFunction);
+      args.push(deps || [], module.nativeTask || taskFunction);
     }
 
     gulp.task.apply(gulp, args);
@@ -66,25 +75,41 @@ function gulpRequireTasks (options) {
     /**
      * Wrapper around user task function.
      * It passes special arguments to the user function according
-     * to the this module configuration.
+     * to the configuration.
      *
      * @param {function} callback
      *
      * @returns {*}
      */
     function taskFunction (callback) {
-      if ('function' === typeof module.fn) {
-        var arguments = _.clone(options.arguments);
-        if (options.passGulp) {
-          arguments.unshift(gulp);
-        }
-        if (options.passCallback) {
-          arguments.push(callback);
-        }
-        return module.fn.apply(module, arguments);
-      } else {
+
+      if ('function' !== typeof module.fn) {
         callback();
+        return;
       }
+
+      let args = [];
+
+      // @deprecated
+      // @todo: remove this in 2.0.0
+      if (options.arguments) {
+        console.warn(
+          'Usage of "arguments" option is deprecated and will be removed in next major version. ' +
+          'Use globals or module imports instead.'
+        );
+        args = Array.from(options.arguments);
+      }
+
+      if (options.passGulp) {
+        args.unshift(gulp);
+      }
+
+      if (options.passCallback) {
+        args.push(callback);
+      }
+
+      return module.fn.apply(module, args);
+
     }
 
     /**
@@ -93,25 +118,30 @@ function gulpRequireTasks (options) {
      * @returns {string}
      */
     function taskNameFromPath (modulePath) {
+
       const relativePath = path.relative(options.path, modulePath);
-      return removeExtension(relativePath)
-        .split(path.sep).join(options.separator)
-      ;
+
+      // Registering root index.js as a default task.
+      if ('index.js' === relativePath) {
+        return 'default';
+      }
+
+      const pathInfo = path.parse(relativePath);
+      const taskNameParts = [];
+
+      if (pathInfo.dir) {
+        taskNameParts.push.apply(taskNameParts, pathInfo.dir.split(path.sep));
+      }
+      if ('index' !== pathInfo.name) {
+        taskNameParts.push(pathInfo.name);
+      }
+
+      return taskNameParts.join(options.separator);
+
     }
 
   }
 
-}
-
-/**
- * Removes extension from the specified path.
- *
- * @param {string} path
- *
- * @returns {string}
- */
-function removeExtension (path) {
-  return path.substr(0, path.lastIndexOf('.'))
 }
 
 /**
@@ -127,7 +157,7 @@ function normalizeModule (module, gulp4) {
       fn: module
     } : {
       fn: module,
-      dep: []
+      deps: []
     };
   } else {
     return module;