+++ CHANGELOG.md
@@ -1,10 +1,35 @@
 # gulp-require-tasks changelog
 
-## Version 1.1.0
-(18 Aug 2016)
+Below is the list of most important changes and versions.
 
-- Support 2 argument syntax for `gulp.task` (gulp 4)
-- *include* and *exclude* options added
+## Version 1.3.0
+(9 Aug 2018)
+
+- Support 2-argument syntax for `gulp.task` (gulp 4)
+- `include` and `exclude` options added
+
+
+## Version 1.2.0
+(30 July 2017)
+
+- Usage of `arguments` option is deprecated, use globals or imports instead
+- Usage of `module.dep` property is deprecated, use `module.deps` instead
+
+
+## Version 1.1.1
+(31 May 2017)
+
+- Fixed backward compatibility with older LTS Node versions
+
+
+## Version 1.1.0
+(16 May 2017)
+
+- Implemented [root directory tasks](README.md#using-root-directory-tasks)
+- `index.js` in the root of tasks directory is now registered as a default task
+- `module.dep` renamed to `module.deps`, `module.dep` is deprecated and will be removed in `2.0`
+- Introduced yarn
+
 
 ## Version 1.0.3
 (16 Mar 2016)